+++ PROP/src/capaDomini/Perfils/PerfilAutor.java
@@ -22,34 +22,34 @@ public class PerfilAutor extends Perfil {
     static ArrayList<String> conferencies = new ArrayList<>();
     static ArrayList<String> articles = new ArrayList<>();
     static ArrayList<String> co_autors = new ArrayList<>();
-    static int quantitat_termes;
-    static int quantitat_conf;
-    static int quantitat_articles;
-    static int quantitat_coautors;
+    static short quantitat_termes;
+    static short quantitat_conf;
+    static short quantitat_articles;
+    static short quantitat_coautors;
     
     
-    public int get_quantitat_termes() {
+    public short get_quantitat_termes() {
         return quantitat_termes;
     }
     
-    public void set_quantitat_termes(int q_terme) {
+    public void set_quantitat_termes(short q_terme) {
         quantitat_termes = q_terme;
     }
     
-    public int get_quantitat_conf() {
+    public short get_quantitat_conf() {
         return quantitat_conf;
     }
-    public void set_quantitat_conf(int q_conferencia) {
+    public void set_quantitat_conf(short q_conferencia) {
         quantitat_conf = q_conferencia;
     }
     
-    public int get_quantitat_articles() {
+    public short get_quantitat_articles() {
         return quantitat_articles;
     }
-    public void set_quantitat_articles(int q_articles) {
+    public void set_quantitat_articles(short q_articles) {
         quantitat_articles = q_articles;
     }
-    public void set_quantitat_autors(int q_coautors) {
+    public void set_quantitat_autors(short q_coautors) {
         quantitat_coautors = q_coautors;
     }
     
@@ -65,7 +65,7 @@ public class PerfilAutor extends Perfil {
             System.out.println(nom);
             System.out.println("");
             System.out.println("Conferencies més rellevants d'aquest autor:");
-            for (int i = 0; i < conferencies.size(); ++i) {
+            for (short i = 0; i < conferencies.size(); ++i) {
                 String aux = conferencies.get(i);
                 System.out.println("\t- "+aux);
             }
@@ -74,7 +74,7 @@ public class PerfilAutor extends Perfil {
             }
             System.out.println("");
             System.out.println("Articles més rellevants d'aquest autor:");
-            for (int i = 0; i < articles.size(); ++i) {
+            for (short i = 0; i < articles.size(); ++i) {
                 String aux = articles.get(i);
                 System.out.println("\t- "+aux);
             }
@@ -83,13 +83,13 @@ public class PerfilAutor extends Perfil {
             }
             System.out.println("");
             System.out.println("Termes més rellevants d'aquest autor:");
-            for (int i = 0; i < termes.size(); ++i) {
+            for (short i = 0; i < termes.size(); ++i) {
                 String aux = termes.get(i);
                 System.out.println("\t- "+aux);
             }
             System.out.println("");
             System.out.println("Co-Autors més rellevants d'aquest autor:");
-            for (int i = 0; i < co_autors.size(); ++i) {
+            for (short i = 0; i < co_autors.size(); ++i) {
                 String aux = co_autors.get(i);
                 System.out.println("\t- "+aux);
             }
